+++ platform_config.yaml
@@ -1,0 +1,18 @@
+description:
+    - "An exemplary robotic lab description. Add devices at will"
+sila_servers:
+    storage:
+        Hotel1:
+            capacity: 20
+        Hotel2:
+            capacity: 20
+    movers:
+        robot_arm:
+            capacity: 1
+
+
+# ------ Translation to used resources in the process description ------
+# the key must match the categories above and the values must match the required resources in the workflow description
+pythonlab_translation:
+    storage: LabwareStorageResource
+    movers: MoverServiceResource
