+++ robot-arm/saved_graphs/position_graph_Dummy.gml
@@ -1,0 +1,21 @@
+graph [
+  node [
+    id 0
+    label "neutral"
+    x 0
+    y 0
+    z 0
+  ]
+  node [
+    id 1
+    label "hotel"
+    x 0
+    y 0
+    z 0
+  ]
+  edge [
+    source 0
+    target 1
+    dist 0.0
+  ]
+]
