+++ robot-arm/saved_graphs/position_graph_XArm.gml
@@ -1,0 +1,57 @@
+graph [
+  node [
+    id 0
+    label "neutral"
+    x 0
+    y 0
+    z 0
+    roll 0
+    pitch 0
+    yaw 0
+  ]
+  node [
+    id 1
+    label "hotel-intermediate"
+    x 0
+    y 0
+    z 0
+    roll 0
+    pitch 0
+    yaw 0
+  ]
+  node [
+    id 2
+    label "hotel1"
+    x 0
+    y 0
+    z 0
+    roll 0
+    pitch 0
+    yaw 0
+  ]
+  node [
+    id 3
+    label "hotel2"
+    x 0
+    y 0
+    z 0
+    roll 0
+    pitch 0
+    yaw 0
+  ]
+  edge [
+    source 0
+    target 1
+    dist 0.0
+  ]
+  edge [
+    source 1
+    target 2
+    dist 0.0
+  ]
+  edge [
+    source 1
+    target 3
+    dist 0.0
+  ]
+]
